// temper/shared-addon-libraries-lib-addon-menu/src/types.ts
export type Valued<T> = T | ((this: void) => T)

export interface LamWidgetData {
  type: string
  name?: Valued<string | number>
  reference?: string
  width?: "full" | "half"
  tooltip?: Valued<string | number>
  tooltipText?: string | number
  warning?: Valued<string | number>
  requiresReload?: boolean
  disabled?: Valued<boolean>
  helpUrl?: Valued<string>
  default?: unknown
  resetFunc?: (this: void, control: LamControl) => void
  getFunc?: (this: void) => unknown
  setFunc?: (this: void, ...args: unknown[]) => void
  registerForRefresh?: boolean
  registerForDefaults?: boolean
  controls?: LamWidgetData[]
}

interface LamValueWidgetData extends LamWidgetData {
  getFunc: (this: void) => unknown
  setFunc: (this: void, ...args: unknown[]) => void
}
// ...
export interface CustomData extends LamWidgetData {
  type: "custom"
  createFunc?: (this: void, control: LamControl) => void
  refreshFunc?: (this: void, control: LamControl) => void
  minHeight?: Valued<number>
  maxHeight?: Valued<number>
}

export interface DescriptionData extends LamWidgetData {
  type: "description"
  text: Valued<string | number>
  title?: Valued<string | number>
  enableLinks?: boolean | ((this: void, ...args: unknown[]) => void)
}

export interface DividerData extends LamWidgetData {
  type: "divider"
  height?: number
  alpha?: number
}

export interface DropdownData extends LamValueWidgetData {
  type: "dropdown"
  name: Valued<string | number>
  choices: string[]
  choicesValues?: unknown[]
  choicesTooltips?: Valued<string | number>[]
  sort?: string
  scrollable?: boolean | number
  multiSelect?: Valued<boolean>
  multiSelectTextFormatter?: Valued<string | number>
  multiSelectNoSelectionText?: Valued<string | number>
  multiSelectMaxSelections?: Valued<number>
}
// ...
export type UpdateValueFn = (this: LamControl, forceDefault?: boolean, ...args: unknown[]) => void
export type UpdateFn = (this: LamControl) => void

export interface IconControl extends TextureControl {
  texture?: string
  tooltip?: string
  color?: ZoColorDef
  size?: number
  OnSelect?: (this: void, icon: Control, texture: string) => void
  SetTexture: (this: IconControl, texture: unknown) => void
}

export interface LamControl extends Control {
  data: LamWidgetData
  panel?: LamControl
  isHalfWidth?: boolean
  container?: Control
  label?: Control
  labelContainer?: Control
  faqControl?: Control
  scroll?: Control
  lineControl?: Control
  startValue?: unknown[]
  value?: unknown
  isDisabled?: boolean
  disabled?: boolean
  disabledLabel?: boolean
  warning?: Control
  info?: Control
  website?: Control
  feedback?: Control
  translation?: Control
  donation?: Control
  controlsToRefresh?: LamControl[]
  icon?: IconControl
  bg?: Control
  arrow?: Control
  animation?: TimelineAnimation
  open?: boolean
  btmToggle?: Control
  button?: Control
  checkbox?: Control
  checkedText?: string
  uncheckedText?: string
  color?: Control
  thumb?: Control
  desc?: Control
  title?: Control
  header?: Control
  divider?: Control
  texture?: Control
  editbox?: Control
  slider?: Control
  minText?: Control
  maxText?: Control
  slidervalueBG?: Control
  slidervalue?: Control
  choices?: Record<string | number, unknown>
  combobox?: Control
  dropdown?: ZoComboBox
  isMultiSelectionEnabled?: boolean
  m_sortType?: unknown
  m_sortOrder?: unknown
  comboboxCount?: number
  dropdownButton?: Control
  UpdateValue?: UpdateValueFn
  UpdateDisabled?: UpdateFn
  UpdateWarning?: UpdateFn
  RefreshPanel?: (this: LamControl) => void
  ForceDefaults?: (this: LamControl) => void
  UpdateChoices?: (this: LamControl, ...args: unknown[]) => void
  SetColor: (this: Control, ...args: unknown[]) => void
  SetIconSize?: (this: LamControl, size: number) => void
  SetDropdownHeight?: (
    this: LamControl,
    dropdown: ZoComboBox,
    dropdownData: DropdownData
  ) => LuaMultiReturn<[number, number, number]>
  AdjustDimensions?: (this: LamControl) => void
}